+++ python/mirage/commands/builtin/github/narrow.py
@@ -18,9 +18,9 @@ from mirage.commands.builtin.github.io import resolve_glob
 from mirage.commands.builtin.grep_helper import (is_literal_pattern,
                                                  search_query)
 from mirage.core.github.constants import SCOPE_WARN
+from mirage.core.github.pushdown import (count_scope_files, scope_relative_key,
+                                         should_use_search)
 from mirage.core.github.repo import ensure_default_branch
-from mirage.core.github.scope import (count_scope_files, scope_relative_key,
-                                      should_use_search)
 from mirage.core.github.search import narrow_paths
 from mirage.core.github.tree import ensure_tree
 from mirage.types import PathSpec
