+++ python/mirage/core/mem0/stat.py
@@ -17,7 +17,7 @@ from typing import Any
 from mirage.accessor.mem0 import Mem0Accessor
 from mirage.cache.index import NULL_INDEX, IndexCacheStore
 from mirage.core.mem0._client import get_memory
-from mirage.core.mem0.scope import ScopeLevel, detect
+from mirage.core.mem0.scope import ScopeLevel, detect_scope
 from mirage.core.render.json import json_bytes
 from mirage.types import FileStat, FileType, PathSpec
 from mirage.utils.errors import enoent
@@ -49,7 +49,7 @@ async def stat(
         path (PathSpec): the path to stat.
         index (IndexCacheStore): index cache.
     """
-    scope = detect(path)
+    scope = detect_scope(path)
     if scope.level == ScopeLevel.ROOT:
         return FileStat(name="/", type=FileType.DIRECTORY)
     if scope.level != ScopeLevel.MEMORY or scope.memory_id is None:
