+++ python/mirage/core/sharepoint/resolve.py
