+++ python/mirage/core/sharepoint/stream.py
@@ -4,8 +4,8 @@ from mirage.accessor.sharepoint import SharePointAccessor
 from mirage.cache.index import NULL_INDEX, IndexCacheStore
 from mirage.core.msgraph.drive_ops import stream_item
 from mirage.core.sharepoint._client import split_path
-from mirage.core.sharepoint._resolver import drive_loc, resolve
 from mirage.core.sharepoint.read import read_bytes
+from mirage.core.sharepoint.resolve import drive_loc, resolve
 from mirage.types import PathSpec
 from mirage.utils.errors import enoent
