+++ python/mirage/core/sharepoint/unlink.py
@@ -2,7 +2,7 @@ from mirage.accessor.sharepoint import SharePointAccessor
 from mirage.cache.context import invalidate_after_unlink
 from mirage.core.sharepoint._client import (GraphError, graph_delete, item_url,
                                             split_path)
-from mirage.core.sharepoint._resolver import resolve
+from mirage.core.sharepoint.resolve import resolve
 from mirage.types import PathSpec
 from mirage.utils.errors import enoent
