+++ python/tests/core/sharepoint/test_find.py
@@ -4,7 +4,7 @@ import pytest
 
 import mirage.core.msgraph.drive_ops as drive_ops
 from mirage.core.sharepoint import find as find_mod
-from mirage.core.sharepoint._resolver import ResolvedPath
+from mirage.core.sharepoint.resolve import ResolvedPath
 from mirage.types import PathSpec
 from mirage.utils.key_prefix import mount_key
