+++ python/tests/core/sharepoint/test_read.py
@@ -2,8 +2,8 @@ import pytest
 from aioresponses import CallbackResult, aioresponses
 
 from mirage.accessor.sharepoint import SharePointAccessor, SharePointConfig
-from mirage.core.sharepoint._resolver import _drive_cache, _site_cache
 from mirage.core.sharepoint.read import read_bytes
+from mirage.core.sharepoint.resolve import _drive_cache, _site_cache
 from mirage.types import PathSpec
 from mirage.utils.key_prefix import mount_key
