+++ typescript/packages/core/src/core/msgraph/_client.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, expect, it, vi } from 'vitest'
 
-import { graphGet, graphList } from './client.ts'
+import { graphGet, graphList } from './_client.ts'
 import { resolveMsGraphConfig } from './config.ts'
 
 afterEach(() => {
