+++ typescript/packages/core/src/core/msgraph/_client.ts
