+++ typescript/packages/core/src/core/onedrive/index.ts
@@ -25,7 +25,7 @@ import type { FindOptions } from '../../resource/base.ts'
 import { FileStat, FileType, type PathSpec } from '../../types.ts'
 import { enoent } from '../../utils/errors.ts'
 import { mountPrefixOf } from '../../utils/key_prefix.ts'
-import { GraphError, graphDelete, graphGet } from '../msgraph/client.ts'
+import { GraphError, graphDelete, graphGet } from '../msgraph/_client.ts'
 import {
   asNumber,
   copyTree,
